+++ Tests/Integration/Convert.Main.Integration.tests.ps1
@@ -320,9 +320,9 @@ try
             WindowsFeatureRule           = $null
             WmiRule                      = $null
         }
-        'U_Windows_10_STIG_V1R12_Manual-xccdf.xml'                    = @{
+        'U_Windows_10_STIG_V1R14_Manual-xccdf.xml'                    = @{
             AccountPolicyRule            = 9
-            AuditPolicyRule              = 38
+            AuditPolicyRule              = 35
             DnsServerRootHintRule        = $null
             DnsServerSettingRule         = $null
             DocumentRule                 = 4
@@ -333,31 +333,7 @@ try
             MimeTypeRule                 = $null
             PermissionRule               = 9
             ProcessMitigationRule        = 31
-            RegistryRule                 = 140
-            SecurityOptionRule           = 6
-            ServiceRule                  = 2
-            SqlScriptQueryRule           = $null
-            UserRightRule                = 30
-            WebAppPoolRule               = $null
-            WebConfigurationPropertyRule = $null
-            WinEventLogRule              = $null
-            WindowsFeatureRule           = 9
-            WmiRule                      = 1
-        }
-        'U_Windows_10_STIG_V1R14_Manual-xccdf.xml'                    = @{
-            AccountPolicyRule            = 9
-            AuditPolicyRule              = 35
-            DnsServerRootHintRule        = $null
-            DnsServerSettingRule         = $null
-            DocumentRule                 = 4
-            FileContentRule              = $null
-            GroupRule                    = 1
-            IisLoggingRule               = $null
-            ManualRule                   = 26
-            MimeTypeRule                 = $null
-            PermissionRule               = 9
-            ProcessMitigationRule        = 31
-            RegistryRule                 = 136
+            RegistryRule                 = 132
             SecurityOptionRule           = 5
             ServiceRule                  = 1
             SqlScriptQueryRule           = $null
@@ -691,9 +667,16 @@ try
                 Should Not Throw
         }
 
+        It 'Should contain Stig Rules' {
+            $output = Get-ChildItem -Path $TestDrive -Filter *.xml
+            [xml] $stigContent = Get-Content -Path $output.FullName -Raw
+
+            $stigContent.DISASTIG | Should -Not -Be $null
+        }
+
         It 'Should append a blank line to the end of the file' {
             $output = Get-ChildItem -Path $TestDrive -Filter *.xml
-            (Get-Content $output.FullName -Raw)[-1] -eq "`n" | Should Be $true
+            (Get-Content -Path $output.FullName -Raw)[-1] -eq "`n" | Should Be $true
         }
     }
     #endregion
@@ -702,4 +685,3 @@ finally
 {
     . $PSScriptRoot\.tests.footer.ps1
 }
-