+++ CHANGELOG.md
@@ -7,7 +7,12 @@
   * Microsoft PowerPoint 2013 Stig V1R6
   * Microsoft Excel 2013 STIG V1R7
   * Microsoft Word 2013 STIG V1R6
+* Added the DSC Composite Resource for Microsoft Office 2013 STIGs
+* Added required module "PolicyFileEditor" for applying Office HKey_Current_User settings
 
+* Updated the following Stigs
+  * WIndows 10 STIG - Updated for xRegistryResource and RegistryRule to Manual Rule
+  
 ## 2.1.0.0
 
 * Migrated Composite resources to the xRegistry resource